fileIsCompressable(): exclude ZFS and use volume capabilities
This is the preferred method to determine if the filesystem hosting a
file supports HFS (decmpfs) compression. Also detect ZFS for Mac
(o3x.org) which can mimic HFS and will then decompress files on the fly.
afsctool will not waste cycles on compression and rewriting files
unless requested with the HFSCOMPRESS_TO_ZFS CMake option (for testing).

https://github.com/RJVB/afsctool/pull/1#discussion_r157676217