first incomplete draft of an equivalent hack for ZFS (very WIP)
The idea is that one can set the dataset compression property to
something different from the current property value and then rewrite
the target files followed by a property reset.
Ideally this would be done at the file-level; with the current lack of
appropriate API one could set the property at the last moment and then
reset it as soon as no workers are active on the dataset in question
anymore. The blunt approach is to set the property when needed and reset
all affected properties after the last file has been rewritten.